+++ src/Data/Text/Short/Encoding/Base64/Lens.hs
@@ -7,7 +7,7 @@
 {-# LANGUAGE Trustworthy #-}
 #endif
 -- |
--- Module       : Data.ByteString.Base64.Lens
+-- Module       : Data.Text.Short.Encoding.Base64.Lens
 -- Copyright 	: (c) 2019-2021 Emily Pillmore
 -- License	: BSD-style
 --
@@ -15,10 +15,10 @@
 -- Stability	: Experimental
 -- Portability	: non-portable
 --
--- This module contains 'Prism''s and 'Iso''s for Base64-encoding and
--- decoding 'ByteString' values.
+-- This module contains 'Prism's and 'Iso''s Base64-encoding and
+-- decoding 'ShortText' values.
 --
-module Data.ByteString.Base64.Lens
+module Data.Text.Short.Encoding.Base64.Lens
 ( -- * Prisms
   _Base64
 , _Base64Url
@@ -33,27 +33,25 @@ module Data.ByteString.Base64.Lens
 , pattern Base64UrlLenient
 ) where
 
-
 import Control.Lens
 
-import Data.ByteString (ByteString)
-import qualified Data.ByteString.Base64 as B64
-import qualified Data.ByteString.Base64.URL as B64U
+import Data.Text.Short (ShortText)
+import qualified Data.Text.Short.Encoding.Base64 as TS64
+import qualified Data.Text.Short.Encoding.Base64.URL as TS64U
 
 
 -- $setup
 --
 -- >>> import Control.Lens
--- >>> import Data.ByteString.Base64.Lens
+-- >>> import Data.Text.Short.Encoding.Base64.Lens
 --
 -- >>> :set -XOverloadedStrings
 -- >>> :set -XTypeApplications
 
-
 -- -------------------------------------------------------------------------- --
 -- Optics
 
--- | A 'Prism'' into the Base64 encoding of a 'ByteString' value
+-- | A 'Prism' into the Base64 encoding of a 'ShortText' value.
 --
 -- >>> _Base64 # "Sun"
 -- "U3Vu"
@@ -61,13 +59,13 @@ import qualified Data.ByteString.Base64.URL as B64U
 -- >>> "U3Vu" ^? _Base64
 -- Just "Sun"
 --
-_Base64 :: Prism' ByteString ByteString
-_Base64 = prism' B64.encodeBase64' $ \s -> case B64.decodeBase64 s of
+_Base64 :: Prism' ShortText ShortText
+_Base64 = prism' TS64.encodeBase64 $ \s -> case TS64.decodeBase64 s of
     Left _ -> Nothing
     Right a -> Just a
 {-# INLINE _Base64 #-}
 
--- | A 'Prism'' into the Base64url encoding of a 'ByteString' value
+-- | A 'Prism' into the Base64-url encoding of a 'ShortText' value.
 --
 -- >>> _Base64Url # "Sun"
 -- "U3Vu"
@@ -75,13 +73,13 @@ _Base64 = prism' B64.encodeBase64' $ \s -> case B64.decodeBase64 s of
 -- >>> "PDw_Pz8-Pg==" ^? _Base64Url
 -- Just "<<???>>"
 --
-_Base64Url :: Prism' ByteString ByteString
-_Base64Url = prism' B64U.encodeBase64' $ \s -> case B64U.decodeBase64 s of
+_Base64Url :: Prism' ShortText ShortText
+_Base64Url = prism' TS64U.encodeBase64 $ \s -> case TS64U.decodeBase64 s of
     Left _ -> Nothing
     Right a -> Just a
 {-# INLINE _Base64Url #-}
 
--- | A 'Prism'' into the Base64url encoding of a 'ByteString' value
+-- | A 'Prism' into the Base64-url encoding of a 'ShortText' value.
 --
 -- Please note that unpadded variants should only be used
 -- when assumptions about the data can be made. In particular, if the length of
@@ -93,17 +91,17 @@ _Base64Url = prism' B64U.encodeBase64' $ \s -> case B64U.decodeBase64 s of
 -- >>> "PDw_Pz4-" ^? _Base64UrlUnpadded
 -- Just "<<??>>"
 --
-_Base64UrlUnpadded :: Prism' ByteString ByteString
-_Base64UrlUnpadded = prism' B64U.encodeBase64Unpadded' $ \s -> case B64U.decodeBase64Unpadded s of
+_Base64UrlUnpadded :: Prism' ShortText ShortText
+_Base64UrlUnpadded = prism' TS64U.encodeBase64Unpadded $ \s -> case TS64U.decodeBase64Unpadded s of
     Left _ -> Nothing
     Right a -> Just a
 {-# INLINE _Base64UrlUnpadded #-}
 
--- | An 'Iso'' into the Base64 encoding of a 'ByteString' value
+-- | An 'Iso'' into the Base64 encoding of a 'ShortText' value
 -- using lenient decoding.
 --
 --
--- _Note:_ This is not a lawful 'Iso'. Please take care!
+-- _Note:_ This is not a lawful 'Iso' in general. Please take care!
 --
 -- >>> _Base64Lenient # "Sun"
 -- "U3Vu"
@@ -111,14 +109,14 @@ _Base64UrlUnpadded = prism' B64U.encodeBase64Unpadded' $ \s -> case B64U.decodeB
 -- >>> "U3Vu" ^. _Base64Lenient
 -- "Sun"
 --
-_Base64Lenient :: Iso' ByteString ByteString
-_Base64Lenient = iso B64.decodeBase64Lenient B64.encodeBase64'
+_Base64Lenient :: Iso' ShortText ShortText
+_Base64Lenient = iso TS64.decodeBase64Lenient TS64.encodeBase64
 
--- | An 'Iso'' into the Base64url encoding of a 'ByteString' value
+-- | An 'Iso'' into the Base64url encoding of a 'ShortText' value
 -- using lenient decoding.
 --
 --
--- _Note:_ This is not a lawful 'Iso'. Please take care!
+-- _Note:_ This is not a lawful 'Iso' in general. Please take care!
 --
 -- >>> _Base64UrlLenient # "<<??>>"
 -- "PDw_Pz4-"
@@ -126,40 +124,40 @@ _Base64Lenient = iso B64.decodeBase64Lenient B64.encodeBase64'
 -- >>> "PDw_Pz4-" ^. _Base64UrlLenient
 -- "<<??>>"
 --
-_Base64UrlLenient :: Iso' ByteString ByteString
-_Base64UrlLenient = iso B64U.decodeBase64Lenient B64U.encodeBase64'
+_Base64UrlLenient :: Iso' ShortText ShortText
+_Base64UrlLenient = iso TS64U.decodeBase64Lenient TS64U.encodeBase64
 
 -- -------------------------------------------------------------------------- --
 -- Patterns
 
--- | Bidirectional pattern synonym for base64-encoded 'ByteString' values.
+-- | Unidirectional pattern synonym for base64-encoded 'ShortText' values.
 --
-pattern Base64 :: ByteString -> ByteString
+pattern Base64 :: ShortText -> ShortText
 pattern Base64 a <- (preview _Base64 -> Just a) where
     Base64 a = _Base64 # a
 
--- | Bidirectional pattern synonym for base64url-encoded 'ByteString' values.
+-- | Unidirectional pattern synonym for base64url-encoded 'ShortText' values.
 --
-pattern Base64Url :: ByteString -> ByteString
+pattern Base64Url :: ShortText -> ShortText
 pattern Base64Url a <- (preview _Base64Url -> Just a) where
     Base64Url a = _Base64Url # a
 
--- | Bidirectional pattern synonym for unpadded base64url-encoded 'ByteString' values.
+-- | Unidirectional pattern synonym for unpadded base64url-encoded 'ShortText' values.
 --
-pattern Base64UrlUnpadded :: ByteString -> ByteString
+pattern Base64UrlUnpadded :: ShortText -> ShortText
 pattern Base64UrlUnpadded a <- (preview _Base64UrlUnpadded -> Just a) where
     Base64UrlUnpadded a = _Base64UrlUnpadded # a
 
--- | Bidirectional pattern synonym for leniently Base64-encoded 'ByteString' values
+-- | Bidirectional pattern synonym for leniently Base64-encoded 'ShortText' values
 --
-pattern Base64Lenient :: ByteString -> ByteString
+pattern Base64Lenient :: ShortText -> ShortText
 pattern Base64Lenient a <- (view (from _Base64Lenient) -> a) where
     Base64Lenient a = view _Base64Lenient a
 {-# COMPLETE Base64Lenient #-}
 
--- | Bidirectional pattern synonym for leniently Base64-encoded 'ByteString' values
+-- | Bidirectional pattern synonym for leniently Base64-encoded 'ShortText' values
 --
-pattern Base64UrlLenient :: ByteString -> ByteString
+pattern Base64UrlLenient :: ShortText -> ShortText
 pattern Base64UrlLenient a <- (view (from _Base64UrlLenient) -> a) where
     Base64UrlLenient a = view _Base64UrlLenient a
 {-# COMPLETE Base64UrlLenient #-}